--- Go/Go_level_2/lesson5/result.txt
@@ -1,12 +1,0 @@
-goos: windows
-goarch: amd64
-cpu: Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz
-BenchmarkTask3Mutex/8-8   	           2856	        436592 ns/op	   33806 B/op	      82 allocs/op
-BenchmarkTask3Mutex/8#01-8         	   10000	    144949 ns/op	    8152 B/op	      20 allocs/op
-BenchmarkTask3Mutex/8#02-8         	   10000	    201766 ns/op	    9420 B/op	      23 allocs/op
-
-BenchmarkTask3RWMutex/8-8          	   12909	     81333 ns/op	    1741 B/op	       6 allocs/op
-BenchmarkTask3RWMutex/8#01-8       	   10000	    273910 ns/op	    2413 B/op	       6 allocs/op
-BenchmarkTask3RWMutex/8#02-8       	   22728	    224973 ns/op	    2232 B/op	       7 allocs/op
-PASS
-ok  	command-line-arguments	19.041s
